Add files via upload
Previously, the person running the addon's linkshell contributions didn't show up in Discord.  This is because linkshell_in and linkshell_out are handled differently.  They both parse and appear in Discord now.

Previsously, Cardian replaced anything between "<" and ">" with "<Discord Content>".  This was to prevent showing user_id in linkshell which Discord supplies when a user is pinged.  It prevented running any FFXI commands using "<" and ">" from processing (<t>, <bt>, <hpp>, etc).  This is fixed.  It can now differentiate between Discord and FFXI "<" / ">" commands.

Addressed a screenshot issue where the file name was incorrect so no action was taken.